+++ legacy/gan/utils.py
