+++ legacy/gan/wgan.py
