+++ workshops/gan/src/training.py
